+++ src/lib-dns/dns-lookup.c
@@ -385,7 +385,7 @@ static int dns_lookup_input_args(struct dns_lookup *lookup, const char *const *a
 static void dns_lookup_save_msecs(struct dns_lookup *lookup)
 {
 	struct timeval now;
-	int diff;
+	long long diff;
 
 	i_gettimeofday(&now);
 
@@ -425,11 +425,11 @@ static int dns_client_input_args(struct connection *conn, const char *const *arg
 
 static void dns_lookup_timeout(struct dns_lookup *lookup)
 {
-	int duration_msecs = timeval_diff_msecs(&ioloop_timeval,
-						&lookup->start_time);
-	lookup->result.error = t_strdup_printf("Lookup timed out in %u.%03u secs",
-					       duration_msecs / 1000,
-					       duration_msecs % 1000);
+	long long duration_msecs = timeval_diff_msecs(&ioloop_timeval,
+						      &lookup->start_time);
+	lookup->result.error =
+		t_strdup_printf("Lookup timed out in %lld.%03lld secs",
+				duration_msecs / 1000, duration_msecs % 1000);
 
 	dns_lookup_callback(lookup);
 	dns_lookup_free(&lookup);
